refactor(tests): move shared helpers into tests package
