--- conftest.py
@@ -1,11 +1,0 @@
-"""Pytest bootstrap for repository-root test collection."""
-from __future__ import annotations
-
-import sys
-from pathlib import Path
-
-ROOT = Path(__file__).resolve().parent
-ROOT_STR = str(ROOT)
-
-if ROOT_STR not in sys.path:
-    sys.path.insert(0, ROOT_STR)
