+++ scripts/slr_lineage.py
@@ -16,7 +16,7 @@ from typing import Mapping
 
 # Dual-path import: when this module is loaded via `from scripts.slr_lineage
 # import emit` from the repo root (the namespace-package style used by
-# test_check_sprint_contract.py, _test_helpers.py, etc.), the sibling import
+# test_check_sprint_contract.py, tests/test_helpers.py, etc.), the sibling import
 # needs the `scripts.` prefix. When loaded via `import slr_lineage` after a
 # sys.path.insert (the style used by scripts/test_slr_lineage_emission.py),
 # the bare name works. Try the package form first; fall back to the sibling
