+++ scripts/test_check_benchmark_report.py
@@ -5,7 +5,7 @@ import unittest
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
-from scripts._test_helpers import run_script
+from tests.test_helpers import run_script
 
 SCRIPT = Path(__file__).resolve().parent / "check_benchmark_report.py"
 SCHEMA = Path(__file__).resolve().parent.parent / "shared" / "benchmark_report.schema.json"
