+++ scripts/test_check_passport_reset_contract.py
@@ -16,7 +16,7 @@ import unittest
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
-from scripts._test_helpers import run_script
+from tests.test_helpers import run_script
 from scripts.check_passport_reset_contract import scan_duplicate_option_values
 
 SCRIPT = Path(__file__).resolve().parent / "check_passport_reset_contract.py"
