+++ scripts/test_check_task_type.py
@@ -6,7 +6,7 @@ import unittest
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
-from scripts._test_helpers import run_skill_linter
+from tests.test_helpers import run_skill_linter
 
 SCRIPT = Path(__file__).resolve().parent / "check_task_type.py"
