+++ scripts/test_validate_compliance_fixtures.py
@@ -3,7 +3,7 @@ import unittest
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
-from scripts._test_helpers import run_script
+from tests.test_helpers import run_script
 
 SCRIPT = Path(__file__).resolve().parent / "validate_compliance_fixtures.py"
 REAL_FIXTURES = Path(__file__).resolve().parent.parent / "examples" / "compliance"
