+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Shared test package for ARS pytest helpers."""
